Implement Normalized Escape Iteration
Updates the core sequence iterator so that it can now (optionally)
support a refinement count, which will produce smooth interpolants
between two successive iteration counts. The result is the ability
to produce much higher quality images at relatively lower iteration
counts, and higher-quality renders in general.

As part of that work, I refactored the mandelbrot sequence logic into
a class to make the code a bit easier to work with.