Improve search objective (#16)
Improves the objective function in `mandelbrot-search` by sampling on a grid over the proposed image, rather than just the center point in the image. This produces images that are centered (more or less) on the interesting features, rather than (essentially) constraining the center of the image to be on the boundary of the set.