Add histogram utility (#17)
Adds a minimalistic histogram utility that can be used to efficiently 
process very large data sets. It also has a method that can be used to
print diagnostics to a file (or the terminal), and a full set of unit
tests. Eventually this will be used for color-mapping the Mandelbrot 
set.