Use Rayon to parallelize Mandelbrot core calculation (#24)
Replaces the simple double-for-loop implementation of the Mandelbrot sequence calculation with a parallel iterator from Rayon, allowing rows of the image to be computed in parallel. The resulting speed up is impressive:

```
cargo run --release --  mandelbrot-render .\examples\mandelbrot_render\default_params.json
```

**serial implementation:**
```
MeasuredElapsedTime:
 -- Setup:      292.4µs
 -- Mandelbrot: 822.7738ms
 -- Histogram:  3.3665ms
 -- CDF:        2.4µs
 -- Color Map:  33.7352ms
 -- Write PNG:  5.8156ms
 ```

**parallel with Rayon:**
```
MeasuredElapsedTime:
 -- Setup:      299.6µs
 -- Mandelbrot: 64.924ms
 -- Histogram:  6.2161ms
 -- CDF:        2.5µs
 -- Color Map:  55.4925ms
 -- Write PNG:  6.5953ms
 ```

This is only a single data point, not a benchmark, but here we're going from ~820 ms down to ~65 ms. This brings us down into the range where you could imagine putting this inside of a render loop that could look smooth. Next "tall pole" is the color map, which can be "compiled" into a look-up table during start-up.