Add datetime argument support (#28)
Refactors the file generation pipeline so that the "date time" naming is optional. 

This makes it easy to run the same command several times and overwrite the results, but then set a flag to generate the output in a stand-alone directory. Includes some other minor clean up as well.

Also adds the "tiny" parameter set, which is useful for debugging quickly.