Add support for "pan" and "zoom" modifications to parameters (#29)
This PR adds the `rescale` and `translate` options to the CLI, allowing the user to iteratively tweak the image produced in "image-relative coordinates", rather than in absolute coordinates. This makes it much easier to adjust things at the command line. Eventually this should be a GUI, but one step at a time.

Here is an example render, tweaked by updating these parameters:
![starfish](https://github.com/MatthewPeterKelly/fractal-renderer/assets/8137529/d6775b6e-72df-40d8-9915-d73b1a423adc)

Here is an example of what this looks like on the command line:
```
 cargo run  --release --  mandelbrot-render .\examples\mandelbrot_render\starfish.json --translate -0.35 -0.42 --rescale 1.8
 ```