Update DDP params with enum for phase angle
Now it is possible to specify a range of phase angles to produce or a
single snapshot. So far, the range is handled by only generating the
first value in the range.

Next step is to update the render function to recursively call itself
(or otherwise) to implement the loop.