Render complete ddp phase cycle (#36)
Adds a new `time_phase.Series` option to the parameters for the DDP fractal, allowing the program to generate a "series" of renders, each with a different initial phase offset. This produces a neat animation that is periodic over [0,1] in the normalized phase offset. The phase offset is in the periodic forcing function at the initial state of the simulation.

The branch between the "snapshot" and "series" modes of the program is handled as an enum, a technique which could work well for implementing a zoom series on the Mandelbrot set: https://github.com/MatthewPeterKelly/fractal-renderer/issues/9.

Here is an example animation, produced by the new sections in the README:
![default_series](https://github.com/MatthewPeterKelly/fractal-renderer/assets/8137529/b0e654d1-2650-480f-934f-ac5666adaae2)