Add anti aliasing for chaos game (#61)
Implements antialiasing for the "chaos game" class of fractals: so far the Barnsley Fern and the Serpinsky Polygons.

The algorithm used here is different than from the DDP and other "directly rendered" fractals. The sampling approach means that we need to "oversample" the image somehow and then filter down that image. The algorithm here is:
- store a "subpixel mask" with each pixel, along with the color
- on each sample, set the pixel color, and then one element from the subpixel mask
- after sampling, scale the pixel color toward the background, using the number of active subpixels
  - if there are only a few subpixels active, then the color will be near the background
  - if all subpixels are active, then the pixel color is unmodified