Refactor CLI:  auto-detect fractal from `.json` (#62)
## Summary

Resolves https://github.com/MatthewPeterKelly/fractal-renderer/issues/58.

Previously, we needed the command line argument to match the parameter `.json` file, and there was no indication from the `.json` file what type of fractal it was for. This made the CLI a bit difficult to use, and also makes it harder to have a directory of examples that span many fractal types (eg. https://github.com/MatthewPeterKelly/fractal-renderer/issues/6).

Now, the `.json` file is self-describing: it includes the name of the fractal that the parameters go to.  Behind the scenes, this is accomplished using an `enum`, which has every type of fractal that we know how to render. That top-level enum is deserialized first, and then a `match` branch is used to select the correct fractal rendering utility.

## Other changes

Most of the lines of code changed in this PR are refactoring the `.json` files, both to improve organization and also to update them to work with the new `render` command. New guidelines:
- use a new auto-formatter (`prettier` JSON extension for VS-Code)
- each example directory is named for the fractal in a consistent way
- use `default` more consistently in examples
- consolidate README to only list examples with "windows" path delimiters

Also rename `render.rs` to `image_utils.rs`.