Add the "full resolution" shadow fern fractal (#74)
This is one of my favorite fractals that I've generated so far. It is the barnsley fern with anti-aliasing turned all of the way up, but then not run fully to convergence. The result is that some of the less-commonly sampled regions show up as "in shadow", which looks neat:

![shadow_fern_QHD](https://github.com/user-attachments/assets/6df42bbb-75e1-4214-9adc-efd49ffe3dec)