Rewrite color map and fix bug (#79)
Rewrites the color map utility to properly support different interpolation types. The interpolation type is handled using generics, so that we can avoid a run-time check on every call to evaluate the color map. This should make it much faster, eg, for using in the GUI, allowing the color map to be recomputed on the fly in response to user input.

Along the way, I found and fixed a bug in the `color_swatch` utility --> now the color swatch is drawn correctly.