Use new color maps in the mandelbrot set (#83)
Removes the hard-coded color maps for the Mandelbrot set and uses the color map utility.

Here is the new render for the top-level mandelbrot view:
![default](https://github.com/user-attachments/assets/f32236ae-d1af-4ca8-af7d-e6a7385bea76)

I think the color maps could still use some tweaks, but this is am improvement. It would probably be good to allow the user to decouple the "color for the points in the mandelbrot set" from a specific color in the color map.