Use `Rgb` as the interface for pixels (#86)
Previously, the "standard interface" for pixels was `{Option<f32>, color map, background color, histogram, CDF}`.... this is too much! Also, the details of exactly how the color is generated are different for each fractal. Instead, move that logic into the fractal-specific code, and then the general utilities all operate directly on pixel buffers.

Now, the interface is just `Rgb<u8>`. This simplfies a few things, and also pushes the histogram and CDF logic into the Mandelbrot set, rather than forcing other fractals to also use those details. One follow-up project that this unblocks is the ability to do the interpoaltion for anti-aliasing in RGB space, rather than in the non-obvious `Option<f32>` space, which required a default value and color map in the loop.

One big side effect of this change is that we do two full passes over the fractal: one for the histogram and one for the fractal itself. This is slower, but I think it is worth it. It also gives us the ability, for example, to compute and cache the histogram (for example in the live GUI version). We can also use a sub-set of the pixels to generate the histogram.