Use natural log output for Mandelbrot set (#87)
**Summary:**

Updates the output of the Mandelbrot set sequence iteration to use "natural log of the normalized iteration count", rather than just the iteration count. This allows us to get high-quality color mapping with much smaller histograms, because the iteration count is logarithmically distributed when the maximum iteration limit is large.

This change is "transparent" when the histogram bin count is on the same order as the maximum iteration count, but makes a large improvement when `bin_count << max_iter_count`.

Also, tweak all of the examples to work nicely with the new scaling. Increase the maximum iteration count across the board.

**Why do this?**

The histogram generation pass is separate from the main render pass. Right now, it is single threaded. I want to make it multi-threaded, so that it can go inside of a fast render loop in the `explore` GUI. The color map looks good with a large number of samples in the histogram, but using a single histogram shared across threads will lead to contention at the lock. The solution, I think, is to have each thread own a histogram, add samples locally, and then join histograms once all samples are collected. This operation will be more efficient if the histogram bins counts are smaller.

Additionally, we now use a look-up table to speed up the rendering step, but it turns out that this look-up table also scales poorly with the maximum iteration count, resulting in a banded appearance unless there are a huge number of entries in the table.