Add support for rendering the Julia set for quadratic maps (#91)
**Summary:**
Add support for rendering the "Julia Set" variants of the same quadratic map equation that underpins the Mandelbrot Set. I implemented it by generalizing the machinery that is used to render the Mandelbrot set. The generalized code lives in `quadratic_map.rs`, with (much shorter) fractal-specific code living in `mandelbrot.rs` and `julia.rs`. 

**Changes:**
- Create a new `ConvergenceParams` struct to collect related parameters
- Add two "julia set" examples
- Fix all of the mandelbrot set examples to support new `ConvergenceParams` struct 
- Generalize the mandelbrot utilities into `quadratic_map.rs`
- Pattern-match the new "julia set" utility from the mandelbrot set version