Add `controller` module for SISO control
Adds the `Controller` and `PointTracker` classes, which implement a
simple scalar tracking controller with a built-in physics simulation.

So far, this only includes unit tests, but eventually this will be used
as the back-end for a "smooth GUI" control, where each parameter that
needs to smoothly change over time will be tied to a point tracker.

The `PointTracker` class collects a controller, state, and target
together with a simulation, allowing the user to request a constant
velocity or a target position. The two modes are both backed by a PD
controller, with some careful gain tuning and limits such that the
velocity never exceeds the reference (velocity mode) or the max speed
(in position target mode).

This took a few tries to get right, so there are some nice unit tests to
show that the various features work as expected.