Generalize `explorer` feature to support the Julia set fractal (#96)
**Summary:** Refactors the "explore" feature to be built on top of the new `RenderWindow` trait, which then is implemented for both the `Julia` and `Mandelbrot` fractals. 

**Details:**
- Also defines a few other traits to make this all work nicely.
- Adds the `PointRenderFn` to reduce writing out complicated return types.

Architecture note:  Implementing this without taking a performance on runtime polymorphism took a few tries. The eventual solution that I came up with was to have the `RenderWindow` trait (used by the `explore` feature) operate based on runtime polymorphism, but then the back-end implementation is fully generic. In practice this means that we incur the extra v-table lookup cost on things like `render_window.pan_view()`, which happens once per image, but not on `generate_scalar_image_in_place()`, which is fully generic (static polymorphism, highly optimized).