Add smooth window control for "explore" mode (#97)
Replaces the "jump the window view on key press or click", to
back-ending the "view port" of the fractal with a simple dynamical
system, where the keyboard and mouse provide acceleration-level commands
to move the window around. The result is a smooth motion. The update
rate is "as fast as the fractal can be rendered", with the underlying
physics implemented to be robust to variable length time steps.

This makes some significant changes to the view control and related
implementation details for the `explore` mode, creating better
interfaces between components.