Enable histogram-backed color map table to update in place (#102)
updates the histogram and color map table code so that they support a "reset in place" method, which allows the maps and tables to be updated without allocating memory. This will help with performance when we switch over the interactive "explore" application to continuously recompute the color map so that we get the right colors as we move around the fractal.