Recompute histogram during `explore` mode
Previously, this histogram was computed once at the start of the
"explore" mode, and then used throughout. The problem was, once you got
to a different region of the fractal, the histogram was wrong and the
color map no longer looked right.

This commit fixes that by continuously updating the histogram during the
exploration. This slows down the frame rate (we'll optimize that in a
follow-up PR), but now the colors look right.

This required a bit more work than expected, as the implementation
required changing a few of the interfaces. The resulting interfaces
should be easier to work with, both in terms of adding new fractal
implementations, and in adding new features to the explore mode, such as
a "fast compute" that could be used while in motion.

The main technical change here was switching to using proper classes
that implement traits for the interfaces, rather than just a simple
lambda function.