Move render calculation to dedicated thread (#106)
* Move render calculation to dedicated thread

Changes the implementation of the `update()` call on the renderable
object to run in a background thread, thus unblocking the `winit`
application from processing user events. This in turn prevents the event
buffer from growing unboundedly when the user holds down a keyboard key.

Now, those events are processed and used to update the view port
continuously, and then whenever the renderer thread is next available it
will get the most recent view port and render that.

* remove useless comment