Use constant speed controller
Replace the "PD control with limits" implementation of the speed
controller for the view window (explore mode) with a constant velocity
controller. This is far simpler, and has the added benefit that it
immediately stops when reaching the target or when the user releases the
key. As a result, we no longer see a visual artifact of small changes in
the image as the controller converges to the target.

The motivation for this change is that I want to add a "lazy render"
check so that the fractal is only computed when the view port or other
parameters change. This check is much more practical if we don't spend a
bunch of time slowly converging to the target view on each user input.