Working PoC
add bilinear itnerpolation

it works!

Works... but scaling laws are not quite there.

Seems like we cap out at 100ms for larger images. Why?

slightly faster?  Nah.

well, that was a bit faster

WIP: experimenting with timing

Looks like the rendering is taking a lot of time...

... but not all of it.

I think it is time to look at the histogram

FOUND IT.

The histogram is SO slow.  Time to parallelize that.

WIP