Parallel histogram generation (#114)
Updates the `populate_histogram()` method to be parallelized across the first dimension of the image. This provides a roughly 10x speed-up in the histogram benchmarks:

**before:**
```
benches/mandelbrot_ice_fracture.json
                        time:   [14.906 ms 14.919 ms 14.933 ms]
benches/mandelbrot_default.json
                        time:   [16.090 ms 16.093 ms 16.096 ms]  
```

**after**
```
benches/mandelbrot_ice_fracture.json
                        time:   [1.4046 ms 1.4178 ms 1.4334 ms]
benches/mandelbrot_default.json
                        time:   [1.5376 ms 1.5494 ms 1.5625 ms]
```

This should help to improve to performance of the "explore mode", which was hitting this step as one of the slowest operations (after the primary render operation, which I plan to speed-up in a following PR by dynamically down-sampling the primary image based on render times).

The implementation works by refactoring the histogram to be backed by a vector of atomic integers, which can then be safely incremented from multiple threads. Then, the histogram is passed to each thread by cloning an `Arc`, rather than using a mutable reference.