Add `downsample_stride` and `RenderOptions` (#115)
Updates many of the fractals to include a `RenderOptions` configuration feature, which allows the fractal to be rendered in blocks. This can be useful, for example, to render at a high frame rate (with low quality) during "explore mode", and then switch to a high-quality (but slow) render once the user has stopped updating the image specification.

By default this is set to one. A future PR will actually use it to do something interesting.