Use bilinear interpolation when down-sampling the rendered image (#123)
Resolves #120.

Getting this working taught me a lot about the borrow checker!

I ended up implementing it in two passes, one of which is fully parallelized with Rayon. The code is pretty well optimized, although you could probably parallize it a bit more using some unsafe code (not worth it for this project though).