Add `explore` support for the Driven-Damped Pendulum attractor (#125)
Adds `explore` support for the Driven-Damped Pendulum attractor fractal, by implementing the required traits on the `DrivenDampedPendulumParams` struct. As a nice side-benefit, that also meant that I could use the standard implementation of `render()` as well, which allowed the custom render method for DDP to be removed.

As part of this PR I removed the ability to generate animations of the DDP attractor across different phase offsets. This is a bummer, but I think it is worth it for (1) clarity of the interface and (2) not breaking the `explore` mode if someone passed the wrong inputs. Part of my motivation here (3) is that I think there is a *much* better way to implement this, and in a way that supports general animation for any fractal that is `Renderable`. I'll create an issue to document the concept --> https://github.com/MatthewPeterKelly/fractal-renderer/issues/126.