WIP: upgrade speed param from int to float
Two changes to speed optimization

(1) "level" is now a floating point value
(2) "level" of zero is "highest quality" and 1.0 is "highest speed"

Fixed the bug. Now it works mostly as before

Next step is to implement a feedback controller on the
actual rate.

WIP: . designing the feedback controller

Probably should split into a "pure root solver" and a domain-specific
wrapper.

formatter passes now

wrote the update function. Shuold work?

TODO:  add a unit test for the update function
- should be continuoue wrt update time
- bounded level on 0,1
- should actually regulate to the specified value

appease clippy, fmt, and test

ihooked it up to the core engine

Correctly set "idle" vs "active" mode

TODO:  stop redrawing once good enough

Fixed the "idle update" logic

Now we should properly cascade back to full resolution

Making a bbetter test.  Last one was a dud.

WIP:  better test, full closed loop system

ifixed render engine... maybe

ijust totally broken. Needs work.

Pause!  Ok, new plan.

This controller is much more like some sort of discrete linear map.
Reminds me a lot of discret linear iteration stuff, i.e. newtons
method and the like. Interestingly... I wonder if we could make a fractal
for rendering the controller itself in state space. He. Anyway...

The output of the controller needs to be "relative" to the previous command.
For example, we want the fixed point to be "if period is in the target window,
then keep the previous command." If the period is too slow, then we need to speed
up, so increase the command. We have a fixed "max command" of 1.0, so we could
get some nice "tuning space" on this by defining an upper bound on acceptable
period. As a bonus, we could also have the render engine "give up" if the
elapsed duration exceeds that boundary.

On the other hand, if we're rendering faster than desired, then we should
reduce the command to improve the render quality. This maps nicely, because
we have both a minimum period (zero) and a minimum command (zero).

In this way, the controller has three regions, which comprise a piecewise
linear function, mapping from "period" to "next command".

Rather than do this piecemeal, we're probably better off just making a
piecewise linear interpoaltion method... or just finding one.