Switch "speed optimization level" to use floating point values on the range [0.0, 1.0] (#138)
Switch "speed optimization level" to use floating point values on the range [0.0, 1.0], which is required to enable adaptive render quality to work properly. That feature will be added as a follow-up. For now, we get the closest thing to the previous functionality working with the new framework. By convention, "zero" is "maximum quality" and "one" is "maximum speed" (at huge reduction in quality). In theory, we should be able to hit any reasonable frame rate at the maximum render speed (by sacrificing quality).

One detail of this change is that all of the "speed optimization levels" are now bounded (by design). This means that each individual fractal and trait can specify the "reasonable range" for that property, and those all get bundled up (hierarchically) into the single top-level speed control value (which eventually will be the output of the regulator).

Once the user has stopped interacting, the interface automatically steps toward an optimization level of zero, so that we leave a high-quality static image on the screen when idle.