Swap interpolation to use pass by move/copy (#139)
The Interpolator trait was passing the lower and upper bounds by reference, which wasn't really necessary -- they data types are always either scalar or small fixed-size vectors, which can be trivially copied. Additionally, in many cases the arguments can be moved into the function, and the reference just added another layer of indirection and likely prevented some compiler-side optimization. The usage of the interface is now a bit simpler as well.