Add support for keyframe modification to the keyframe interpolator (#140)
Adds two new methods to the `KeyframeInterpolator`, enabling the user to modify both the query and value of a keyframe. For now, this feature is just covered by tests, but in a follow-up PR this will be used for adaptive frame rate regulation.

Also fix some naming inconsistency and extend test coverage for the two new methods.