Working!  Better. Still needs improvement.
A few issues:

(1) We're still chattering a lot on the max command
step, which indicates that our controller model is just bad.
Maybe that is fine, but if so, perhaps we should just get
rid of it entirely and just do the stepping controller.
Another option to avoid the chattering might be to just lowpass
filter the output of the controller and be done with it.

(2) there are a ton of text printouts still, and those
need to be cleaned up and the removed. It would be really
nice to have a signals logging frameworks...