Add heuristic map render policy (#145)
Builds on top of #143 to add a policy that maps from the measured render period to the render quality. This enables the explore mode to regulate the frame-rate, providing a more-responsive interface to the user. Once the user stops interacting with the keyboard, the quality is gradually increased, eventually (after a few frames) reaching maximum render quality (as specified by the user configs).