Use standard cargo example format for examples. (#146)
Large refactor of the entire `examples/` directory to use standard
cargo example format. Additionally, the render and explore variants of
each example are separate now, enabling the exact details to be tweaked
as needed.

In addition to updating the examples, this commit also updates the unit
tests that do validation of the parameter files used in the examples.