Upgrade the "speed optimizer" library utilities (#150)
Fixes a minor bug in the scaling laws for the DDP fractal (tolerance was going the wrong way). Now it can actually be used effectively in explore mode!

Rewrite the library utilities for scaling parameters so that they are easier to use correctly, and also natively support both linear and logarithmic scaling. Changes the behavior slightly to be easier to understand and use for new fractals.  Adds a ton of unit tests for the library code.